livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.10,<3.13"
dependencies = [
    "chromadb>=1.4.1",
    "langchain>=1.2.9",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.8",
    "livekit-agents[hedra,images,liveavatar,silero,turn-detector]~=1.3.6",
    "livekit-plugins-cartesia>=1.3.12",
    "livekit-plugins-langchain~=1.1",
    "livekit-plugins-liveavatar>=1.3.12",
    "livekit-plugins-noise-cancellation~=0.2",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
]
